refactor(back): move Callback and CallbackResponse types
